requirements_dev.txt
# commands
invoke==2.2.0

# pre-commit
pre-commit==3.5.0

# testing
pytest==7.4.3
coverage==7.6.9

# linting
ruff==0.1.3

# docs
mkdocs==1.5.3
mkdocs-material==9.4.6
mkdocstrings-python==1.12.2

# configuration
hydra-core==1.3.2

# profiling
torch_tb_profiler==0.4.3
pyprof2calltree==1.4.7